# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NutriUDF API",
    "description": "API para a plataforma SaaS de gestão nutricional NutriUDF",
    "version": "1.0.0",
    "contact": { "name": "NutriUDF", "email": "suporte@nutriudf.com" }
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "Desenvolvimento" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Patient": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Maria Silva" },
          "email": { "type": "string", "format": "email" },
          "age": { "type": "integer", "example": 30 },
          "sex": { "type": "string", "enum": ["M", "F", "O"] },
          "weight": { "type": "number", "example": 65.5 },
          "height": { "type": "number", "example": 1.65 },
          "imc": { "type": "string", "example": "24.07" },
          "goal": { "type": "string", "example": "Emagrecimento" },
          "notes": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Consultation": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "patient_id": { "type": "string", "format": "uuid" },
          "starts_at": { "type": "string", "format": "date-time" },
          "ends_at": { "type": "string", "format": "date-time" },
          "title": { "type": "string" },
          "notes": { "type": "string" },
          "status": { "type": "string", "enum": ["SCHEDULED", "COMPLETED", "CANCELLED", "NO_SHOW"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/patients": {
      "get": {
        "tags": ["Patients"],
        "summary": "Listar pacientes do nutricionista",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Lista de pacientes", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Patient" } } } } },
          "401": { "description": "Não autenticado" }
        }
      },
      "post": {
        "tags": ["Patients"],
        "summary": "Criar paciente",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Patient" } } } },
        "responses": {
          "201": { "description": "Paciente criado" },
          "400": { "description": "Dados inválidos" }
        }
      }
    },
...
    "/consultations": {
      "get": {
        "tags": ["Consultations"],
        "summary": "Listar consultas",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Lista de consultas" } }
      },
      "post": {
        "tags": ["Consultations"],
        "summary": "Agendar consulta",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Consultation" } } } },
        "responses": { "201": { "description": "Consulta criada" }, "400": { "description": "Dados inválidos" } }
      }
    },
...
    "/plans": {
      "post": {
        "tags": ["Plans"],
        "summary": "Criar plano alimentar",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "patientId": { "type": "string" }, "planDate": { "type": "string", "format": "date" }, "mealData": { "type": "object" } } } } } },
        "responses": { "201": { "description": "Plano criado" } }
      }
    },
...
    "/foods": {
      "get": {
        "tags": ["Foods"],
        "summary": "Buscar alimentos",
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string" }, "description": "Termo de busca (nome ou código TBCA)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": { "200": { "description": "Lista de alimentos com macros e micros" } }
      }
    },
...
  }
}